// src/types/generated/variantResponseDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Inventory Management API
 * A comprehensive NestJS inventory and order management system with role-based access control, stock tracking, and audit logging.
 * OpenAPI spec version: 1.0
 */
import type { ProductInventoryDto } from './productInventoryDto';
import type { VariantResponseDtoBarcode } from './variantResponseDtoBarcode';
import type { VariantResponseDtoImageUrl } from './variantResponseDtoImageUrl';

export interface VariantResponseDto {
  id: string;
  sku: string;
  name: string;
  costPrice: number;
  sellingPrice: number;
  isActive: boolean;
  /** @nullable */
  imageUrl: VariantResponseDtoImageUrl;
  /** @nullable */
  barcode: VariantResponseDtoBarcode;
  inventory: ProductInventoryDto | null;
}